Use separate dto package
