--- app/src/main/java/com/book/library/book/AnyBookDto.java
@@ -1,3 +1,0 @@
-package com.book.library.book;
-
-public record AnyBookDto(int id, String name, String author, boolean borrowed) implements BookDto {}
