--- app/src/main/java/com/book/library/book/BookDto.java
@@ -1,3 +1,0 @@
-package com.book.library.book;
-
-sealed interface BookDto permits AnyBookDto, AvailableBookDto, BorrowedBookDto {}
