+++ app/src/main/java/com/book/library/book/BookRepository.java
@@ -1,5 +1,6 @@
 package com.book.library.book;
 
+import com.book.library.dto.AvailableBookDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,7 +10,7 @@ public interface BookRepository extends JpaRepository<Book, Integer> {
 
     @Query(
             """
-            SELECT new com.book.library.book.AvailableBookDto(b.id, b.name, b.author)
+            SELECT new com.book.library.dto.AvailableBookDto(b.id, b.name, b.author)
             FROM Book b
             LEFT JOIN b.borrowed bb
             WHERE bb IS NULL""")
