+++ app/src/main/java/com/book/library/book/BookService.java
@@ -1,5 +1,8 @@
 package com.book.library.book;
 
+import com.book.library.dto.AnyBookDto;
+import com.book.library.dto.AvailableBookDto;
+import com.book.library.dto.BorrowedBookDto;
 import com.book.library.reader.Reader;
 import com.book.library.reader.ReaderDto;
 import com.book.library.reader.ReaderRepository;
