+++ app/src/main/java/com/book/library/book/BorrowedBookRepository.java
@@ -1,5 +1,6 @@
 package com.book.library.book;
 
+import com.book.library.dto.BorrowedBookDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,7 +10,7 @@ public interface BorrowedBookRepository extends JpaRepository<BorrowedBook, Inte
 
     @Query(
             """
-            SELECT new com.book.library.book.BorrowedBookDto(bb.id, bb.borrowedOn, b.name, b.author, r.firstName, r.lastName)
+            SELECT new com.book.library.dto.BorrowedBookDto(bb.id, bb.borrowedOn, b.name, b.author, r.firstName, r.lastName)
             FROM BorrowedBook bb
             JOIN bb.book b
             JOIN bb.reader r""")
