+++ app/src/main/java/com/book/library/dto/AnyBookDto.java
@@ -1,0 +1,3 @@
+package com.book.library.dto;
+
+public record AnyBookDto(int id, String name, String author, boolean borrowed) implements DataTransferObject {}
