+++ app/src/main/java/com/book/library/dto/DataTransferObject.java
@@ -1,0 +1,3 @@
+package com.book.library.dto;
+
+public sealed interface DataTransferObject permits AnyBookDto, AvailableBookDto, BorrowedBookDto {}
