+++ app/src/test/java/com/book/library/book/BorrowedBookRepositoryTest.java
@@ -1,5 +1,7 @@
 package com.book.library.book;
 
+import com.book.library.dto.AvailableBookDto;
+import com.book.library.dto.BorrowedBookDto;
 import com.book.library.reader.Reader;
 import com.book.library.reader.ReaderRepository;
 import com.book.library.repository.BaseRepositoryTest;
